T459: Extract shared getActivePhases to _gsd-helpers.js — DRY fix
Duplicate getActivePhases function in gsd-branch-gate and gsd-pr-gate
extracted to _gsd-helpers.js (prefixed _ so load-modules skips it).
30/30 GSD gate tests passing.
