+++ modules/PreToolUse/gsd-pr-gate.js
@@ -6,37 +6,7 @@
 "use strict";
 var fs = require("fs");
 var path = require("path");
-
-/**
- * Parse active phase numbers from ROADMAP.md
- */
-function getActivePhases(projectDir) {
-  var roadmap = path.join(projectDir, ".planning", "ROADMAP.md");
-  if (!fs.existsSync(roadmap)) return [];
-
-  try {
-    var content = fs.readFileSync(roadmap, "utf-8");
-    var phases = [];
-    var inActive = false;
-    var lines = content.split("\n");
-
-    for (var i = 0; i < lines.length; i++) {
-      var line = lines[i];
-      if (/^##\s+Active Milestone/i.test(line)) {
-        inActive = true;
-        continue;
-      }
-      if (inActive && /^##\s/.test(line) && !/^###/.test(line)) break;
-      if (inActive) {
-        var phaseMatch = line.match(/^###\s+Phase\s+(\d+)/i);
-        if (phaseMatch) phases.push(phaseMatch[1]);
-      }
-    }
-    return phases;
-  } catch (e) {
-    return [];
-  }
-}
+var getActivePhases = require("./_gsd-helpers").getActivePhases;
 
 module.exports = function(input) {
   if (input.tool_name !== "Bash") return null;
